Fix problem w/ YUML generator.
The test had been disabled because of perceived errors in the yuml service.  This, however, was not the case.  If you are fetching the YUML directly via requests,
you need to have vertical bars in the URL.  If, however, you are recording the YUML in a markdown document, you need to escape the vertical bars (`&#124;`)