Change the default default_range message to a logger info to clean up the output
Changed message that a slot_usage slot was created to an info as there are often cases where this is what one wants to do
Set the default range on a newly defined slot_range.  Note: One might want to check whether there are other tweaks that are
done to slots earlier on that should be applied to a non-inherited usage as well.