Final fix for issue #168
Get the remainder of an old pull request associated with this issue incorporated.
A couple of additional tweaks on the emit_metadata entry.
Added narrow and broad mappings to the core metamodel (Version 1.5.4)
Fixed a testing problem where output was pre-filtered, meaning that we didn't have gen date and the like
Turned RDF compare on now that the revision seems to be ok
Turned off the 3.7 unit test on a pull request, as it gets tested elsewhere